% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_and_simulation_generic_functions.R
\name{simulekde}
\alias{simulekde}
\title{Simulate observations and return kernel density estimates}
\usage{
simulekde(model, y0, nbreps = 300)
}
\arguments{
\item{model:}{a model object}

\item{N:}{integer, for population size}

\item{nbreps:}{number of replicates}
}
\description{
Simulate observations and return kernel density estimates
}
\examples{
popmodelfunction = model.Pareto.bernstrat
theta=4;xi=1;conditionalto=list(N=10000,sampleparam=list(tauh=c(0.01,0.1)))
model<-popmodelfunction(theta,xi,conditionalto);y0=1.5
Obs<-generate.observations(model)
simulekde(model,1,100)
library(ggplot2)
ggplot(data=data.frame(y=Obs$y), aes(y)) +geom_density()+stat_function(fun=model$dloiy)
}

