% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2_kde.R
\name{varfHT0}
\alias{varfHT0}
\title{popmodelfunction = model.Pareto.bernstrat
theta=4;xi=1;conditionalto=list(N=10000,sampleparam=list(tauh=c(0.01,0.1)))
model<-popmodelfunction(theta,xi,conditionalto);y0=1.5
.Obs<-generate.observations(model);
.ker=ker;
.h=ks::hpi(x=.Obs$y)
varfHT0(y0,Obs)}
\usage{
varfHT0(y0, .Obs, .ker = ker, .h = ks::hpi(x = .Obs$y))
}
\description{
popmodelfunction = model.Pareto.bernstrat
theta=4;xi=1;conditionalto=list(N=10000,sampleparam=list(tauh=c(0.01,0.1)))
model<-popmodelfunction(theta,xi,conditionalto);y0=1.5
.Obs<-generate.observations(model);
.ker=ker;
.h=ks::hpi(x=.Obs$y)
varfHT0(y0,Obs)
}

